refactor(ir): accept any relation in `ops.ExistsSubquery`
