+++ ibis/expr/operations/__init__.py
@@ -15,6 +15,7 @@ from ibis.expr.operations.relations import *  # noqa: F403
 from ibis.expr.operations.sortkeys import *  # noqa: F403
 from ibis.expr.operations.strings import *  # noqa: F403
 from ibis.expr.operations.structs import *  # noqa: F403
+from ibis.expr.operations.subqueries import *  # noqa: F403
 from ibis.expr.operations.temporal import *  # noqa: F403
 from ibis.expr.operations.temporal_windows import *  # noqa: F403
 from ibis.expr.operations.udf import *  # noqa: F403
