+++ ibis/expr/types/logical.py
@@ -313,7 +313,7 @@ class BooleanColumn(NumericColumn, BooleanValue):
         def resolve_exists_subquery(outer):
             """An exists subquery whose outer leaf table is unknown."""
             (inner,) = (t for t in parents if t != outer.op())
-            relation = ops.Project(ops.Filter(inner, [self]), {"1": 1})
+            relation = ops.Filter(inner, [self])
             return ops.ExistsSubquery(relation).to_expr()
 
         if len(parents) == 2:
